refactor: dynamic github link save file
